Address set not used and implicit conversion

+++ larpandoracontent/LArReclustering/LArExample/RandomFigureOfMeritTool.h
@@ -46,7 +46,7 @@ private:
 
 inline float RandomFigureOfMeritTool::GetRandomFom()
 {
-    return static_cast<float>(rand()) / RAND_MAX;
+    return static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
 }
 
 } // namespace lar_content
